+++ src/core/ChatSDKErrors.ts
@@ -1,0 +1,5 @@
+enum ChatSDKErrors {
+    UnsupportedPlatform = "UnsupportedPlatform"
+}
+
+export default ChatSDKErrors;