+++ src/telemetry/TelemetryEvent.ts
@@ -22,6 +22,7 @@ enum TelemetryEvent {
     GetLiveChatTranscript = "GetLiveChatTranscript",
     CreateIC3Adapter = "CreateChatAdapter",
     CreateACSAdapter = "CreateACSAdapter",
+    CreateDirectLine = "CreateDirectLine",
     GetVoiceVideoCalling = "GetVoiceVideoCalling",
     GetIC3Client = "GetIC3Client",
     InitializeVoiceVideoCallingSDK = "InitializeVoiceVideoCallingSDK",